* YAM_MA.c, mui/ReadMailGroup.c, mui/WriteWindow.c, tcp/pop3.c: fixed some warnings

+++ src/mui/ReadMailGroup.c
@@ -997,7 +997,7 @@ DECLARE(ReadMail) // struct Mail *mail, ULONG flags
       {
         char numStr[SIZE_SMALL];
 
-        snprintf(numStr, sizeof(numStr), "%d", xget(rmData->readWindow, MUIA_ReadWindow_Num));
+        snprintf(numStr, sizeof(numStr), "%d", (int)xget(rmData->readWindow, MUIA_ReadWindow_Num));
         MA_StartMacro(MACRO_READ, numStr);
       }
       else
